+++ src/frame_cache.h
@@ -27,7 +27,7 @@ void vcache_destroy(void **p);
 void vcache_resize(void **p, int size);
 void vcache_clear (void *p);
 
-uint8_t *vcache_get_buffer(void *p, void *dc, int id, int64_t frame, int w, int h, int fmt, void **cptr);
+uint8_t *vcache_get_buffer(void *p, void *dc, unsigned short id, int64_t frame, short w, short h, int fmt, void **cptr);
 void vcache_release_buffer(void *p, void *cptr);
 
 size_t vcache_info_html(void *p, char *m, size_t n); // write HTML to m - max length n
